feat: add subscriptions
prompt: Add a new button for signed in users. This button should be below the RSS button. Clicking on it should take the user to a new subscribe page registered as a route at /. On this page the users should see a list of products. Next to each product should be a checkbox telling the user whether or not they are "subscribed" to that product. THe user should be able to check or uncheck any boxes and click "save". The data for this subscription should be located at /users/$id/subscriptions/ and should contain the ID of the product. Saving should update the firebase data, and the initial state should be loaded from firebase.